# docs/swagger.yaml
basePath: /
definitions:
  main.Data:
    type: object
  main.Response:
    properties:
      error:
        type: string
      result: {}
      success:
        type: boolean
    type: object
host: nftinfo.online
info:
  contact:
    email: seup87@@gmail.com
    name: API Support
    url: https://github.com/Jeongseup
  description: This is a sample server.
  termsOfService: http://swagger.io/terms/
  title: Celestia DA API
  version: "1.0"
paths:
  /{hash}:
    get:
      description: Retrieves a blob from Celestia DA by commitment hash
      parameters:
      - description: Commitment hash
        in: path
        name: hash
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response'
      summary: Retrieves blob by commitment hash
      tags:
      - Celestia DA
  /{namespace}/{index_number}:
    get:
      description: Retrieves a blob from Celestia DA by namespace key and index number
      parameters:
      - description: Namespace key
        in: path
        name: namespace
        required: true
        type: string
      - description: Index number
        in: path
        name: index_number
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response'
      summary: Retrieves blob by namespace key and index number
      tags:
      - Celestia DA
  /node_info:
    get:
      description: Pings Celestia DA node and returns node info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response'
      summary: Returns Celestia DA node info
      tags:
      - Celestia DA
  /retrieve_blob:
    get:
      description: Retrieves a blob from Celestia DA by height and namespace key
      parameters:
      - description: Block height
        in: query
        name: height
        required: true
        type: integer
      - description: Namespace key
        in: query
        name: namespace_key
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response'
      summary: Retrieves blob by height and namespace key
      tags:
      - Celestia DA
  /submit_formdata:
    post:
      description: Submits form data (image) to Celestia DA
      parameters:
      - description: Image file
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response'
      summary: Submits form data
      tags:
      - Celestia DA
  /submit_metadata:
    post:
      description: Submits JSON data to Celestia DA
      parameters:
      - description: Data payload
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/main.Data'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.Response'
      summary: Submits JSON data
      tags:
      - Celestia DA
swagger: "2.0"